+++ app/main/__init__.py
@@ -24,6 +24,7 @@ from app.main.views import (  # noqa
     jobs,
     join_service,
     letter_branding,
+    make_your_service_live,
     manage_users,
     new_password,
     notifications,
