+++ tests/app/main/views/service_settings/test_letter_branding_requests.py
@@ -307,7 +307,7 @@ def test_POST_letter_branding_request_creates_zendesk_ticket(
     org_id,
     expected_org_name,
 ):
-    mock_send_zendesk = mocker.patch("app.main.views.service_settings.index.zendesk_client.send_ticket_to_zendesk")
+    mock_send_zendesk = mocker.patch("app.main.views.service_settings.branding.zendesk_client.send_ticket_to_zendesk")
     service_one["organisation"] = org_id
 
     client_request.post(
