+++ tests/conftest.py
@@ -4484,3 +4484,14 @@ def mock_get_notifications_count_for_service(mocker):
 @pytest.fixture(scope="function")
 def fake_nonce():
     return "TESTs5Vr8v3jgRYLoQuVwA"
+
+
+@pytest.fixture
+def mock_get_service_settings_page_common(
+    mock_get_all_letter_branding,
+    mock_get_inbound_number_for_service,
+    mock_get_free_sms_fragment_limit,
+    mock_get_service_data_retention,
+    mock_get_organisation,
+):
+    return
